Adding system for showing consolidated score

+++ db/models.py
@@ -139,7 +139,7 @@ class Attachment(models.Model):
 class Analysis(models.Model):
     notes = models.TextField()
     source = models.CharField(max_length=50)
-    score = models.IntegerField()
+    score = models.DecimalField(max_digits=3, decimal_places=2)
     # one to many relationship between an email and analysis, respectively
     email = models.ForeignKey('Email')
     first_seen = models.DateTimeField(editable=False)
